Separate text-only and vision variants of the same model architecture
